Revert "yauheni/74383/ converting app.jsx in account to functional component (#6282)"
This reverts commit e1fc3e8af2bc0bfefa3427e0bcf257300563755c.
