+++ packages/account/src/index.js
@@ -3,7 +3,7 @@ import { makeLazyLoader, moduleLoader } from '@deriv/shared';
 import { Loading } from '@deriv/components';
 
 const App = makeLazyLoader(
-    () => moduleLoader(() => import(/* webpackChunkName: "account-app", webpackPreload: true */ './App')),
+    () => moduleLoader(() => import(/* webpackChunkName: "account-app", webpackPreload: true */ './App.jsx')),
     () => <Loading />
 )();
